Refactored the loading of children to a separate file to keep each component as clean as possible
